fix(rms_norm): cast row_idx to int64 in _rms_norm_forward_kernel(# 803)

+++ src/liger_kernel/ops/rms_norm.py
@@ -63,7 +63,7 @@ def _rms_norm_forward_kernel(
     3. https://arxiv.org/pdf/1910.07467
     """
 
-    row_idx = tl.program_id(0)
+    row_idx = tl.program_id(0).to(tl.int64)
     col_offsets = tl.arange(0, BLOCK_SIZE)
     mask = col_offsets < n_cols
